Commentimpl

+++ src/main/java/com/avanzada/alojamientos/DTO/comment/CommentDTO.java
@@ -2,11 +2,13 @@ package com.avanzada.alojamientos.DTO.comment;
 
 import com.avanzada.alojamientos.DTO.host.ReplyHostDTO;
 
+
+
 import java.time.LocalDateTime;
 
 public record   CommentDTO (
         Long id,
-        Integer rating,
+        Float rating,
         String text,
         LocalDateTime createdAt,
         Boolean isModerated,
